dom/battery/test/test_battery_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Battery API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Battery API **/

ok('battery' in navigator, "navigator.battery should exist");

var battery = navigator.battery;
ok(battery.level >= 0.0 && battery.level <= 1.0, "Battery level " + battery.level + " should be in the range [0.0, 1.0]");

if (battery.charging) {
  ok(battery.chargingTime >= 0, "Battery chargingTime " + battery.chargingTime + " should be nonnegative when charging");
  is(battery.dischargingTime, Infinity, "Battery dischargingTime should be Infinity when charging");
} else {
  is(battery.chargingTime, Infinity, "Battery chargingTime should be Infinity when discharging");
  ok(battery.dischargingTime > 0, "Battery dischargingTime " + battery.dischargingTime + " should be positive when discharging");
}

</script>
</pre>
</body>
</html>
